split WlSeat and related code out from wayland_connector.cpp
