+++ src/server/frontend/wayland/CMakeLists.txt
@@ -14,6 +14,7 @@ set(
   output_manager.cpp            output_manager.h
   wl_mir_window.cpp             wl_mir_window.h
   wl_surface.cpp                wl_surface.h
+  wl_seat.cpp                   wl_seat.h
 )
 
 add_library(
